M-INT-Claude/requirements.txt
fastapi
uvicorn
openai
pydantic
pillow
piexif
opencv-python
imagehash
python-multipart
requests
